+++ packages/widgetbook/lib/src/fields/field_type.dart
@@ -39,4 +39,7 @@ enum FieldType {
 
   /// For fields that have a [Object] value, represented as a segmented button.
   objectSegmented,
+
+  /// For fields that have a [Iterable] value, represented as a segmented button.
+  iterableSegmented,
 }