+++ packages/widgetbook_annotation/CHANGELOG.md
@@ -1,4 +1,4 @@
-## Unreleased
+## 3.8.0
 
 - **FEAT**: Allow excluding use-cases with `@UseCase.exclude`. ([#1676](https://github.com/widgetbook/widgetbook/pull/1676) - by [@EArminjon](https://github.com/EArminjon))
 - **FEAT**: Add `IterableKnobConfig` to support the new `iterable` knob. ([#1678](https://github.com/widgetbook/widgetbook/pull/1678) - by [@EArminjon](https://github.com/EArminjon))